+++ bootstrap.ts
@@ -17,12 +17,6 @@ function log(msg) {
   Zotero.logError(msg)
 }
 
-// In Zotero 6, bootstrap methods are called before Zotero is initialized, and using include.js
-// to get the Zotero XPCOM service would risk breaking Zotero startup. Instead, wait for the main
-// Zotero window to open and get the Zotero object from there.
-//
-// In Zotero 7, bootstrap methods are not called until Zotero is initialized, and the 'Zotero' is
-// automatically made available.
 async function waitForZotero() {
   if (typeof Zotero != 'undefined') {
     await Zotero.initializationPromise
