+++ client/config.js
