+++ client/en-US/scite-zotero-plugin.properties
