--- client/install.rdf
@@ -1,33 +1,0 @@
-<?xml version="1.0"?>
-<RDF xmlns="http://www.w3.org/1999/02/22-rdf-syntax-ns#"
-        xmlns:em="http://www.mozilla.org/2004/em-rdf#">
-
-    <Description about="urn:mozilla:install-manifest">
-        <em:id>.hi@scite.ai.</em:id>
-        <em:name>Scite Plugin Zotero (V7)</em:name>
-        <em:version>1.1</em:version>
-        <em:multiprocessCompatible>true</em:multiprocessCompatible>
-        <em:updateURL>https://zotero-download.s3.amazonaws.com/tmp/scite-zotero-plugin/updates-1.1.json</em:updateURL>
-        <em:homepageURL>https://github.com/zotero/make-it-red</em:homepageURL>
-        <em:localized>
-            <Description>
-                <em:locale>en-US</em:locale>
-                <em:name>Scite Plugin Zotero (V7) </em:name>
-                <em:description>Scite Plugin Zotero (V7)</em:description>
-            </Description>
-        </em:localized>
-
-        <em:creator>Zotero</em:creator>
-        <em:type>2</em:type> <!-- type=extension -->
-        <em:iconURL>chrome://scite-zotero-plugin/skin/icon.png</em:iconURL>
-        <em:icon64URL>chrome://scite-zotero-plugin/skin/icon64.png</em:icon64URL>
-
-        <em:targetApplication>
-            <Description>
-                <em:id>zotero@chnm.gmu.edu</em:id>
-                <em:minVersion>6.999</em:minVersion>
-                <em:maxVersion>*</em:maxVersion>
-            </Description>
-        </em:targetApplication>
-    </Description>
-</RDF>
