+++ client/skin/contrasting.png
